Address IndexedDB key-value review feedback
