+++ .changeset/brave-keys-commit.md
@@ -1,0 +1,5 @@
+---
+"@effect/platform-browser": patch
+---
+
+Fix IndexedDB-backed key-value writes to wait for transaction commit before reporting success.
